+++ man/NEWS.md
@@ -26,3 +26,7 @@
 - Changed from rjson to jsonlite
 - Cleaned up examples
 - parameter argument is now a list and does not need to be converted to JSON
+
+# rosetteApi 1.7.0
+- New name_deduplication endpoint
+- New name_deduplication example